SmallAutoField

+++ tests/model_inheritance/test_abstract_inheritance.py
@@ -331,21 +331,21 @@ class AbstractInheritanceTests(TestCase):
 
         self.assertEqual(
             fields(model1),
-            [('id', models.AutoField), ('name', models.CharField), ('age', models.IntegerField)]
+            [('id', models.BigAutoField), ('name', models.CharField), ('age', models.IntegerField)]
         )
 
-        self.assertEqual(fields(model2), [('id', models.AutoField), ('name', models.CharField)])
+        self.assertEqual(fields(model2), [('id', models.BigAutoField), ('name', models.CharField)])
         self.assertEqual(getattr(model2, 'age'), 2)
 
-        self.assertEqual(fields(model3), [('id', models.AutoField), ('name', models.CharField)])
+        self.assertEqual(fields(model3), [('id', models.BigAutoField), ('name', models.CharField)])
 
-        self.assertEqual(fields(model4), [('id', models.AutoField), ('name', models.CharField)])
+        self.assertEqual(fields(model4), [('id', models.BigAutoField), ('name', models.CharField)])
         self.assertEqual(getattr(model4, 'age'), 2)
 
         self.assertEqual(
             fields(model5),
             [
-                ('id', models.AutoField), ('foo', models.IntegerField),
+                ('id', models.BigAutoField), ('foo', models.IntegerField),
                 ('concretemodel_ptr', models.OneToOneField),
                 ('age', models.SmallIntegerField), ('concretemodel2_ptr', models.OneToOneField),
                 ('name', models.CharField),
